Derive DCAuthManager and AdAuthManager from BaseADAuthmanager
